backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
python-decouple==3.8
